Move stuff to a .parse subpackage so we can start extracting inner classes a bit without too much mess
